# Index DefType tag keys to speed up GetMetadataBy*Key calls