During game loading, WeaponDefHandler.cpp constructs all known WeaponDefs. Each WeaponDef object is constructed by finding the [lua table key <-> WeaponDef member variable mapping] in DefType::tagMetaData. This search is done by linearly scanning DefType::tagMetaData.

The construction process' exection time grows in proportion to (number of tags) * (number of weapons).

To cut down on loading times, we ensure that DefType::Load calls uses an index to look up tags.

Additionally, this change makes GetMetaDataByInternalKey look up keys case-insensitively.

- There is no change to the DEFTAG and DUMMYTAG API macros.
- Slightly reduces static storage required.
- Slightly increases the pre-main dynamic initialization time.
- Significantly speeds up DefType::Load and DefType::ReportUnknownTags calls.

Tested against BAR - there are no differences in `DefType::GetMetaDataByInternalKey` and `DefType::GetMetaDataByExternalKey` calls.

## Further Details

The fundamental source of the slowness lies in `DefType::GetMetaDataByExternalKey` and `DefType::GetMetaDataByInternalKey` which both linearly scans the `tagMetaData` array for a tag matching the specified key. Both of these GetMetaDataBy*Key calls are used repeatedly in `DefType::Load` calls through the chain `DefType::Load` -> `defInitFuncs`, defined in the `DEFTAG_CNTER` macro -> `DefType::GetTag` as well as in `DefType::ReportUnknownTags`.

The solution to this problem is to create an index from `tag name string` to `DefTagMetaData*`.

A first pass implementation of the solution puts the construction of the index in `DefType::AddTag`. However, the fluent interface exposed by `DefTagBuilder` allows for modification of member `DefTagBuilder::data` after the call to `DefType::AddTag` (see the FIXME in DefType::AddTag). Specifically, this means that properties such as `externalName` and `fallbackName` can be changed after the initial `DefType::AddTag` call. There are two potential solutions to this issue:

- [Chosen solution] Implicitly finalize the `DefTagBuilder` once the fluent chain is complete. In most fluent interface this would require a `.build()` call at the end of the chain, but that would require changes in WeaponDef.cpp. Here, we rely on the `BuiltDefTag` type's constructor to automatically finalize the `DefTagBuilder` by calling `DefTagBuilder::AddTag`.
- Alternatively, we could construct the index on demand. I.e. we could put a `std::call_once(.., /* construct tagMetaDataBy*Name indices */)` in DefTag::Load such that the first DefTag::Load call populates the map.


## Profiler Screenshots

### Before
~580ms to initialize all WeaponDefs (in `WeaponDefHandler::Init`)
~200us in pre-main dynamic initialization of the various WeaponDef tags.
![before_objdef](https://github.com/beyond-all-reason/spring/assets/2730943/6813f8ca-8289-494f-86c4-e7648de7ad7e)

### After
~100ms to initialize all WeaponDefs (in `WeaponDefHandler::Init`)
~250us in pre-main dynamic initialization of various WeaponDef tags. (Not included in the screenshot is about another ~10us in AllocateTag calls in pre-main initialization).
![after_objdef](https://github.com/beyond-all-reason/spring/assets/2730943/8c2f87ca-0928-49ac-930e-db99226bb4da)