@lhog 2024-01-11
I failed to comprehend `DefinitionTag.h` fully, but looks good from the first glance.