@lhog 2024-01-12
Yeah I know my comment was not criticize your work in any way, it's just that the file is a bit hard to read. But that was the case before your changes too. Your additions are fine.